feat(ppoll): support timeout future

+++ NoAxiom/lib/console/src/lib.rs
@@ -16,7 +16,7 @@ impl Stdout {
 impl Write for Stdout {
     fn write_str(&mut self, s: &str) -> fmt::Result {
         for c in s.chars() {
-            platform::putchar(c as usize);
+            platform::putchar(c as u8);
         }
         Ok(())
     }
