+++ NoAxiom/lib/platform/src/basedev/riscv64_qemu.rs
@@ -4,14 +4,14 @@ pub struct Base;
 
 impl BaseFu for Base {
     #[inline]
-    fn putchar(ch: usize) {
-        sbi_rt::legacy::console_putchar(ch);
+    fn putchar(ch: u8) {
+        sbi_rt::legacy::console_putchar(ch as usize);
     }
 
     /// read a byte, return -1 if nothing exists.
     #[inline]
-    fn getchar() -> usize {
-        sbi_rt::legacy::console_getchar()
+    fn getchar() -> u8 {
+        sbi_rt::legacy::console_getchar() as u8
     }
 
     #[inline]
